Moved all JSON schema rendering code to `NautobotMetadata`
- Only the Metadata class emits JSON schema now when an `OPTIONS` request is made
- Legacy `NautobotMetadata` renamed to `NautobotMetadataV1` for now
- All other code that was in `nautobot.core.api.views` has been ripped out and there is no longer a need to overload or modify anything with REST API views
- Made `eslint` happy
