+++ nautobot/ui/src/views/generic/ObjectCreate.js
@@ -1,5 +1,5 @@
 import { Card, CardHeader, CardBody } from "@chakra-ui/react"; // TODO import from nautobot-ui when available
-import { Button, Frame, Text } from "@nautobot/nautobot-ui";
+import { Button, Frame } from "@nautobot/nautobot-ui";
 import Form from "@rjsf/chakra-ui";
 import validator from "@rjsf/validator-ajv8";
 import axios from "axios";
